Refactor per PR feedback
